+++ server/controllers/studentController.js
@@ -2,9 +2,13 @@ const { Student } = require("../db/models");
 
 class StudentController {
     async getAll(req, res) {
-        const students = await Student.findAll();
+        try {
+            const students = await Student.findAll();
 
-        return res.json(students);
+            return res.json(students);
+        } catch (error) {
+            return res.sendStatus(500);
+        }
     }
 
     async create(req, res) {
@@ -12,17 +16,33 @@ class StudentController {
             ...req.body,
         };
 
-        const result = await Student.create(student);
+        try {
+            const result = await Student.create(student);
 
-        return res.json(result.dataValues);
+            return res.status(201).json(result);
+        } catch (error) {
+            return res.sendStatus(500);
+        }
     }
 
     async delete(req, res) {
         const { id } = req.params;
 
-        await Student.destroy({ where: { id: id } });
+        if (isNaN(id)) {
+            return res.sendStatus(400);
+        }
 
-        return res.json();
+        if (Student.findOne({ where: { id: id } }) == null) {
+            return res.sendStatus(404);
+        }
+
+        try {
+            await Student.destroy({ where: { id: id } });
+
+            return res.sendStatus(204);
+        } catch (error) {
+            return res.sendStatus(500);
+        }
     }
 }
